+++ projects/ontimize-web-ngx-charts-quickstart/src/app/main/bar-chart/multi-bar-horizontal/multi-bar-horizontal.component.ts
@@ -1,0 +1,79 @@
+﻿import { Component, ViewChild } from '@angular/core';
+import { MultiBarHorizontalChartConfiguration, OChartComponent } from 'ontimize-web-ngx-charts';
+
+const BASIC_USAGE_HTML_DATA = `
+<o-chart type="multiBarHorizontalChart" [data]="data" x-label="Time" y-label="Amount (€)" x-data-type="time" [color]="colorScheme" show-x-axis="false">
+`;
+
+@Component({
+  selector: 'multi-bar-horizontal',
+  templateUrl: './multi-bar-horizontal.component.html'
+})
+export class MultiBarHorizontalComponent {
+
+  @ViewChild('multiBarHor')
+  protected multiBarHor: OChartComponent;
+
+  chartParameters: MultiBarHorizontalChartConfiguration;
+  colorScheme = {
+    domain: ['#eeeeee', '#8ab2d2', '#c5c5c5']
+  };
+  data = [
+    {
+      "name": "Germany",
+      "value": 40632,
+      "extra": {
+        "code": "de"
+      }
+    },
+    {
+      "name": "United States",
+      "value": 50000,
+      "extra": {
+        "code": "us"
+      }
+    },
+    {
+      "name": "France",
+      "value": 36745,
+      "extra": {
+        "code": "fr"
+      }
+    },
+    {
+      "name": "United Kingdom",
+      "value": 36240,
+      "extra": {
+        "code": "uk"
+      }
+    },
+    {
+      "name": "Spain",
+      "value": 33000,
+      "extra": {
+        "code": "es"
+      }
+    },
+    {
+      "name": "Italy",
+      "value": 35800,
+      "extra": {
+        "code": "it"
+      }
+    }
+  ];
+
+  getBasicUsageFiles() {
+    return {
+      'html': {
+        'data': BASIC_USAGE_HTML_DATA
+      },
+      'scss': {
+        'data': ''
+      },
+      'typescript': {
+        'data': ''
+      }
+    }
+  }
+}