+++ projects/ontimize-web-ngx-charts-quickstart/src/app/shared/app.services.config.ts
@@ -1,0 +1,6 @@
+﻿
+export const SERVICE_CONFIG = {
+  'customers': {
+    'path': '/customers'
+  }
+};