+++ projects/ontimize-web-ngx-charts/src/lib/components/chart-on-demand/o-chart-on-demand.module.ts
@@ -1,29 +1,20 @@
-import { CommonModule } from '@angular/common';
 import { Injector, NgModule } from '@angular/core';
-import { FormsModule } from '@angular/forms';
-import { O_CHART_ON_DEMAND_SERVICE, OntimizePreferencesService, OntimizeWebModule, preferencesServiceFactory } from 'ontimize-web-ngx';
+import { O_CHART_ON_DEMAND_SERVICE, OntimizePreferencesService, preferencesServiceFactory } from 'ontimize-web-ngx';
 
 import { OChartOnDemandService } from '../../services/chart-on-demand.service';
-import { OChartComponentModule } from '../chart/o-chart.module';
 import { LoadPreferencesDialogComponent } from './load-preferences-dialog/load-preferences-dialog.component';
 import { OChartOnDemandComponent } from './o-chart-on-demand.component';
 import { SavePreferencesDialogComponent } from './save-preferences-dialog/save-preferences-dialog.component';
 
-
+/**
+ * @deprecated Use `OChartOnDemandComponent` directly (standalone). This module will be removed in a future version.
+ */
 @NgModule({
-  declarations: [OChartOnDemandComponent, SavePreferencesDialogComponent, LoadPreferencesDialogComponent],
-  imports: [
-    CommonModule,
-    OntimizeWebModule,
-    FormsModule,
-    OChartComponentModule,
-  ],
+  imports: [OChartOnDemandComponent, SavePreferencesDialogComponent, LoadPreferencesDialogComponent],
   exports: [OChartOnDemandComponent],
   providers: [
     { provide: O_CHART_ON_DEMAND_SERVICE, useClass: OChartOnDemandService },
     { provide: OntimizePreferencesService, useFactory: preferencesServiceFactory, deps: [Injector] }
   ]
 })
-export class OChartOnDemandComponentModule {
-  constructor() { }
-}
+export class OChartOnDemandComponentModule { }