+++ projects/ontimize-web-ngx-charts/src/lib/components/chart-on-demand/save-preferences-dialog/save-preferences-dialog.component.ts
@@ -1,22 +1,25 @@
 import { Component, Injector } from '@angular/core';
-import { UntypedFormGroup, UntypedFormControl, Validators } from '@angular/forms';
-import { MatDialogRef } from '@angular/material/dialog';
-import { OTableBaseDialogClass } from 'ontimize-web-ngx';
+import { ReactiveFormsModule, FormGroup, FormControl, Validators } from '@angular/forms';
+import { MatButtonModule } from '@angular/material/button';
+import { MatDialogModule, MatDialogRef } from '@angular/material/dialog';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatInputModule } from '@angular/material/input';
+import { OntimizeWebModule, OTableBaseDialogClass } from 'ontimize-web-ngx';
 
 @Component({
+  standalone: true,
   selector: 'app-save-preferences-dialog',
-  templateUrl: './save-preferences-dialog.component.html'
+  templateUrl: './save-preferences-dialog.component.html',
+  imports: [MatDialogModule, MatFormFieldModule, MatInputModule, MatButtonModule, ReactiveFormsModule, OntimizeWebModule]
 })
 
 export class SavePreferencesDialogComponent extends OTableBaseDialogClass {
   public name: string;
   public description: string;
 
-  public formGroup: UntypedFormGroup = new UntypedFormGroup({
-    name: new UntypedFormControl('', [
-      Validators.required
-    ]),
-    description: new UntypedFormControl('')
+  public formGroup = new FormGroup({
+    name: new FormControl<string>('', { validators: [Validators.required], nonNullable: true }),
+    description: new FormControl<string>('', { nonNullable: true })
   });
 
   constructor(